+++ crates/plugin_resolve/src/resolver/utils.rs
@@ -1,10 +1,6 @@
 use farmfe_core::{common::PackageJsonInfo, farm_profile_function, regex, serde_json::Value};
 use once_cell::sync::Lazy;
-use std::{
-  path::{PathBuf},
-  str::FromStr,
-};
-
+use std::{path::PathBuf, str::FromStr};
 
 static PACKAGE_REGEX: Lazy<regex::Regex> =
   Lazy::new(|| regex::Regex::new(r"^(?P<group1>[^@][^/]*)|^(?P<group2>@[^/]+/[^/]+)").unwrap());