+++ examples/react-ssr/src/types.d.ts
@@ -1,0 +1,4 @@
+declare module "*.png" {
+  declare const v: string;
+  export default v;
+}