+++ examples/tree-shake-antd/e2e.spec.ts
@@ -2,6 +2,7 @@ import { test, expect, describe } from 'vitest';
 import { startProjectAndTest } from '../../e2e/vitestSetup';
 import { basename, dirname } from 'path';
 import { fileURLToPath } from 'url';
+import { execa } from 'execa';
 
 const name = basename(import.meta.url);
 const projectPath = dirname(fileURLToPath(import.meta.url));
@@ -22,6 +23,10 @@ describe(`e2e tests - ${name}`, async () => {
       },
       command
     );
+
+  await execa('npm', ['run', 'build'], {
+    cwd: projectPath,
+  })
 
   test(`exmaples ${name} run start`, async () => {
     await runTest();