+++ packages/core/src/config/types.ts
@@ -168,6 +168,7 @@ export interface ResolvedCompilation
   resolve?: {
     dedupe?: never;
   } & Config['config']['resolve'];
+  assets?: Omit<Config['config']['assets'], 'mode'>;
   css?: ResolvedCss;
 }