+++ rust-plugins/sass/tests/mod.rs
@@ -3,7 +3,8 @@ use std::{collections::HashMap, fs, io::Write, path::PathBuf, sync::Arc};
 use farmfe_compiler::Compiler;
 use farmfe_core::{
   config::{
-    bool_or_obj::BoolOrObj, preset_env::PresetEnvConfig, AliasItem, Config, ResolveConfig, RuntimeConfig, SourcemapConfig, StringOrRegex
+    bool_or_obj::BoolOrObj, preset_env::PresetEnvConfig, AliasItem, Config, ResolveConfig,
+    RuntimeConfig, SourcemapConfig, StringOrRegex,
   },
   context::CompilationContext,
   module::ModuleType,
@@ -86,12 +87,10 @@ fn test_with_compiler() {
       tree_shaking: Box::new(BoolOrObj::Bool(false)),
       progress: false,
       resolve: Box::new(ResolveConfig {
-        alias: vec![
-          AliasItem::Complex {
-            find: StringOrRegex::String("@".to_string()),
-            replacement: cwd.to_string_lossy().to_string(),
-          },
-        ],
+        alias: vec![AliasItem::Complex {
+          find: StringOrRegex::String("@".to_string()),
+          replacement: cwd.to_string_lossy().to_string(),
+        }],
         ..Default::default()
       }),
       ..Default::default()