+++ scripts/bump-nightly.mjs
@@ -1,23 +1,7 @@
 import { execSync } from "child_process";
 
-// generate nightly version
-const dateString =
-  new Date().getFullYear() +
-  String(new Date().getMonth() + 1).padStart(2, "0") +
-  String(new Date().getDate()).padStart(2, "0");
-
-const gitHash = execSync("git rev-parse --short HEAD").toString().trim();
-
-const nightlyVersion = `2.0.0-nightly.${dateString}.${gitHash}`;
-
-execSync(`npx changeset version --snapshot nightly --no-snap-release`, {
-  stdio: "inherit",
-});
-
-execSync(`pnpm --filter "../packages/cli" --filter "../packages/core" version ${nightlyVersion} --no-git-tag-version`, {
+execSync(`npx changeset version --snapshot nightly`, {
   stdio: "inherit",
 });
 
 execSync("pnpm install --no-frozen-lockfile", { stdio: "inherit" });
-
-console.log(`Nightly version bump completed. Version: ${nightlyVersion}`);