+++ js-plugins/postcss/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @farmfe/js-plugin-postcss
 
+## 2.0.0-nightly-20241023020505
+
+### Patch Changes
+
+- Updated dependencies [3651244]
+  - @farmfe/core@2.0.0-nightly-20241023020505
+
+## 2.0.0-nightly-20241022143924
+
+### Patch Changes
+
+- Updated dependencies [83838bd]
+  - @farmfe/core@2.0.0-nightly-20241022143924
+
+## 2.0.0-nightly-20241022124925
+
+### Patch Changes
+
+- @farmfe/core@2.0.0-nightly-20241022124925
+
+## 2.0.0-nightly-20241022041556
+
+### Patch Changes
+
+- Updated dependencies [24347c9]
+  - @farmfe/core@2.0.0-nightly-20241022041556
+
+## 2.0.0-nightly-20241022014521
+
+### Patch Changes
+
+- Updated dependencies [a82a5b2]
+  - @farmfe/core@2.0.0-nightly-20241022014521
+
 ## 1.8.1
 
 ### Patch Changes