+++ packages/runtime-plugin-import-meta/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @farmfe/runtime-plugin-import-meta
 
+## 1.0.0-nightly-20241023020505
+
+### Major Changes
+
+- cab79e8: bump major version with runtime
+
 ## 0.2.2
 
 ### Patch Changes